fix(quantization): address review feedback on weight FQ calibration change
- Move enable_weight_fake_quant/disable_activation_fake_quant into a new
  _fake_quant_utils module so they can import FakeQuantizeImplBase and
  CompressionTargetTensor at module scope instead of via local imports,
  keeping _utils.py free of dependencies on other quantization modules.
- Use basic_config (weight + activation quantization) instead of
  input_activation_only_config in test_calibration_mode so the weight-FQ
  branch is actually exercised; only assert scale drift for activation
  modules, since weight ranges are fixed at prepare time.
- Drop the changelog.d entry.
